requirements.txt
langgraph>=0.2
langchain-core>=0.2
langchain-google-genai>=1.0
python-dotenv>=1.0
